[FEATURE IMPL] Upload a simple example demonstrate how to use Micrium message queue
I just create a simple example that has two task, Ex_MainStartTask which has a higher priority, and AppGraphics_Task has a lower priority. The AppGraphics_Task just wait for the message to be sent to the message queue, and then toggle LED1, the Ex_MainStartTask will send a message to a queue with 500ms interval.